pii-detector-backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pdf2image
Pillow
# fpdf
PyPDF2
python-docx
reportlab
numpy
opencv-python